Conflicts:
	EPMS.Implementation/Services/DepartmentService.cs
	EPMS.Implementation/Services/JobTitleService.cs
	EPMS.Interfaces/IServices/IJobTitleService.cs
	EPMS.Interfaces/Repository/IEmployeeRepository.cs
	EPMS.Web/Areas/HR/Controllers/DepartmentController.cs
	EPMS.Web/Areas/HR/Controllers/JobTitleController.cs
	EPMS.Web/Areas/HR/Views/Department/Index.cshtml
	EPMS.Web/EPMS.Web.csproj

+++ EPMS.Interfaces/IServices/IDepartmentService.cs
@@ -13,7 +13,7 @@ namespace EPMS.Interfaces.IServices
         IEnumerable<Department> GetAll();
         DepartmentResponse GetAllDepartment(DepartmentSearchRequest departmentSearchRequest);
 
-        Department FindDepartmentById(int? id);
+        Department FindDepartmentById(long? id);
 
         bool AddDepartment(Department department);
         bool UpdateDepartment(Department department);
